fix: compare native amplitude against float64 oracle

+++ src/eval/native_granularity_amplitude_conformance.py
@@ -147,8 +147,8 @@ def reference_outputs(
         ]
     )
     return (
-        np.ascontiguousarray(density, dtype=np.float32),
-        np.ascontiguousarray(sigma, dtype=np.float32),
+        np.ascontiguousarray(density, dtype=np.float64),
+        np.ascontiguousarray(sigma, dtype=np.float64),
     )
